+++ agabpylib/stats/__init__.py
@@ -1,0 +1,7 @@
+"""
+Statistical functions.
+
+Anthony Brown June 2019 - June 2019
+"""
+
+from .distributions import *