+++ agabpylib/stats/distributions.py
@@ -1,0 +1,9 @@
+#
+# Anthony Brown June 2019 - June 2019
+#
+
+from . import _continuous_distns
+
+from ._continuous_distns import *
+
+__all__ = _continuous_distns._distn_names